Move pool and client stats into util/stats
