+++ lib/dispatcher/client.js
@@ -4,10 +4,10 @@ const assert = require('node:assert')
 const net = require('node:net')
 const http = require('node:http')
 const util = require('../core/util.js')
+const { ClientStats } = require('../util/stats.js')
 const { channels } = require('../core/diagnostics.js')
 const Request = require('../core/request.js')
 const DispatcherBase = require('./dispatcher-base')
-const ClientStats = require('./client-stats')
 const {
   InvalidArgumentError,
   InformationalError,
